machines/first/mcp/pyproject.toml
[project]
name = "first-mcp"
version = "0.1.0"
description = "MCP server for the first machine"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "fastmcp>=2.14.2",
    "nats-py>=2.12.0",
    "puda-drivers",
]

[tool.uv.sources]
puda-drivers = {workspace = true}
